confluence-qna-bot/requirements.txt
langchain
openai
faiss-cpu
tqdm
python-dotenv
streamlit
requests
beautifulsoup4
atlassian-python-api
langchain_community